fix(api): fixed test issues
