+++ apps/api/src/app/notifications/dtos/activities-response.dto.ts
@@ -1,6 +1,8 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import {
   DaysEnum,
+  DigestTypeEnum,
+  DigestUnitEnum,
   ExecutionDetailsSourceEnum,
   ExecutionDetailsStatusEnum,
   MessageTemplateDto,
@@ -14,7 +16,59 @@ import {
 } from '@novu/shared';
 import { IsArray, IsEnum, IsNumber, IsOptional, IsString } from 'class-validator';
 import { StepFilterDto } from '../../shared/dtos/step-filter-dto';
-import { DigestMetadataDto } from './digest-metadata.dto';
+
+export class DigestMetadataDto {
+  @ApiPropertyOptional({ description: 'Optional key for the digest' })
+  digestKey?: string;
+
+  @ApiPropertyOptional({ description: 'Amount for the digest', type: Number })
+  amount?: number;
+
+  @ApiPropertyOptional({ description: 'Unit of the digest', enum: DigestUnitEnum })
+  unit?: DigestUnitEnum;
+
+  @ApiProperty({
+    enum: [...Object.values(DigestTypeEnum)],
+    enumName: 'DigestTypeEnum',
+    description: 'The Digest Type',
+    type: String,
+  })
+  type: DigestTypeEnum;
+
+  @ApiPropertyOptional({
+    type: 'array',
+    items: {
+      type: 'object',
+      additionalProperties: true,
+    },
+    description: 'Optional array of events associated with the digest, represented as key-value pairs',
+  })
+  events?: Record<string, unknown>[];
+
+  // Properties for Regular Digest
+  @ApiPropertyOptional({
+    description: 'Regular digest: Indicates if backoff is enabled for the regular digest',
+    type: Boolean,
+  })
+  backoff?: boolean;
+
+  @ApiPropertyOptional({ description: 'Regular digest: Amount for backoff', type: Number })
+  backoffAmount?: number;
+
+  @ApiPropertyOptional({
+    description: 'Regular digest: Unit for backoff',
+    enum: [...Object.values(DigestUnitEnum)],
+    enumName: 'DigestUnitEnum',
+  })
+  backoffUnit?: DigestUnitEnum;
+
+  @ApiPropertyOptional({ description: 'Regular digest: Indicates if the digest should update', type: Boolean })
+  updateMode?: boolean;
+
+  // Properties for Timed Digest
+  @ApiPropertyOptional({ description: 'Configuration for timed digest', type: () => DigestTimedConfigDto })
+  timed?: DigestTimedConfigDto;
+}
 
 export class ActivityNotificationStepResponseDto {
   @ApiProperty({ description: 'Unique identifier of the step', type: String })
